+++ crates/nix_rs/src/lib.rs
@@ -8,4 +8,5 @@ pub mod env;
 pub mod flake;
 pub mod info;
 pub mod refs;
+pub mod store;
 pub mod version;