+++ crates/nix_rs/CHANGELOG.md
@@ -5,6 +5,7 @@
 - **`flake::url`**
   - Add `without_attr`, `get_attr`
   - Simplify the return type of `RootQualifiedAttr::eval_flake`
+- **`store`**: Add module (upstreamed from nixci)
 
 ## 1.0.0